refactor(m2): REF-201 新建 @pixuli/core 包
从 pixuli-common 迁出 types、platform、operation-log、纯 utils、
应用级 locales 与 StorageProvider 插件契约；common 保留兼容 re-export。

Closes #60
