+++ packages/common/src/services/index.ts
@@ -1,14 +1,16 @@
 export { GiteeStorageService } from './giteeStorageService';
 export { GitHubStorageService } from './githubStorageService';
-export { DefaultPlatformAdapter } from './platformAdapter';
-export type { PlatformAdapter } from './platformAdapter';
+export {
+  DefaultPlatformAdapter,
+  type PlatformAdapter,
+} from '@pixuli/core/platform';
 export { logInterceptorService } from './logInterceptorService';
 export type { LogEntry, LogLevel } from './logInterceptorService';
-export { OperationLogService } from './operationLog';
-export type {
-  IOperationLogStorage,
-  OperationLogServiceOptions,
-} from './operationLog';
+export {
+  OperationLogService,
+  type IOperationLogStorage,
+  type OperationLogServiceOptions,
+} from '@pixuli/core/operation-log';
 
 // 图片处理（Web/Desktop 纯 Web 实现，依赖 Canvas；移动端勿用）
 export {
