+++ packages/core/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@pixuli/core",
+  "version": "1.0.0",
+  "description": "Pixuli core — shared types, platform adapters, utils, and plugin contracts",
+  "type": "module",
+  "license": "MIT",
+  "exports": {
+    ".": "./src/index.ts",
+    "./types": "./src/types/index.ts",
+    "./plugins": "./src/plugins/index.ts",
+    "./platform": "./src/platform/index.ts",
+    "./utils": "./src/utils/index.ts",
+    "./locales": "./src/locales/index.ts",
+    "./operation-log": "./src/operation-log/index.ts"
+  },
+  "scripts": {
+    "test": "vitest run",
+    "test:watch": "vitest"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0",
+    "jsdom": "^25.0.1",
+    "typescript": "^5.0.0",
+    "vitest": "^2.1.8"
+  }
+}
