+++ packages/core/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @pixuli/core — 三端共享业务内核（无 UI）
+ */
+export * from './types';
+export * from './plugins';
+export * from './platform';
+export * from './operation-log';
+export * from './utils';
+export * from './locales';
