+++ packages/core/src/locales/app/en-US.json
