+++ packages/core/src/locales/app/zh-CN.json
