+++ packages/core/src/platform/__tests__/platformAdapter.test.ts
