+++ packages/core/src/platform/index.ts
@@ -1,0 +1,4 @@
+export {
+  DefaultPlatformAdapter,
+  type PlatformAdapter,
+} from './platformAdapter';
