+++ packages/core/src/platform/platformAdapter.ts
