+++ packages/core/src/plugins/index.ts
@@ -1,0 +1,5 @@
+export * from './types';
+export {
+  createStoragePluginRegistry,
+  DefaultStoragePluginRegistry,
+} from './registry';
