+++ packages/core/src/types/global.d.ts
