+++ packages/core/src/types/image.ts
