+++ packages/core/src/utils/__tests__/imageUtils.test.ts
