+++ packages/core/src/utils/fileSizeUtils.ts
@@ -1,0 +1,21 @@
+/**
+ * 文件大小格式化工具函数
+ */
+
+/**
+ * 将字节数转换为人类可读的文件大小字符串
+ * @param bytes 字节数
+ * @param decimals 小数位数，默认为2
+ * @returns 格式化后的文件大小字符串
+ */
+export function formatFileSize(bytes: number, decimals: number = 2): string {
+  if (bytes === 0) return '0 B';
+
+  const k = 1024;
+  const dm = decimals < 0 ? 0 : decimals;
+  const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
+}
