+++ vitest.workspace.ts
@@ -1,6 +1,7 @@
 import { defineWorkspace } from 'vitest/config';
 
 export default defineWorkspace([
+  './packages/core/vitest.config.ts',
   // packages/common - 需要 jsdom 环境
   './packages/common/vitest.config.ts',
   // apps/pixuli - jsdom 环境
